+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/InstallerModule.java
@@ -24,6 +24,8 @@ import java.util.Set;
  * First search provided packages for registered installers
  * {@link FeatureInstaller}. Then scan classpath obe more time
  * with installers to apply extensions.
+ * <p>If installer implements {@link BindingInstaller} then it's bind method will be called just after
+ * extension type registration ({@code bind(extType)}), allowing more complicated bindings.</p>
  * <p>Feature installers can be disabled from bundle config.</p>
  * NOTE: Classpath scan will load all found classes in configured packages. Try to reduce scan scope
  * as much as possible.
@@ -143,6 +145,9 @@ public class InstallerModule extends AbstractModule {
                         FeatureUtils.getInstallerExtName(installer.getClass()), type.getName());
                 holder.register(installer.getClass(), type);
                 binder().bind(type);
+                if (installer instanceof BindingInstaller) {
+                    ((BindingInstaller) installer).bind(binder(), type);
+                }
             }
         }
         return recognized;