+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/feature/DummyPlugin1.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.dropwizard.guice.support.feature
+
+import ru.vyarus.dropwizard.guice.module.installer.feature.plugin.Plugin
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 09.10.2014
+ */
+@Plugin(PluginInterface)
+class DummyPlugin1 implements PluginInterface {
+}