+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/feature/PluginInterface.groovy
@@ -1,0 +1,10 @@
+package ru.vyarus.dropwizard.guice.support.feature
+
+/**
+ * Sample plugin interface.
+ *
+ * @author Vyacheslav Rusakov 
+ * @since 09.10.2014
+ */
+interface PluginInterface {
+}