+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/feature/abstr/AbstractPlugin.groovy
@@ -1,0 +1,12 @@
+package ru.vyarus.dropwizard.guice.support.feature.abstr
+
+import ru.vyarus.dropwizard.guice.module.installer.feature.plugin.Plugin
+import ru.vyarus.dropwizard.guice.support.feature.PluginInterface
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 09.10.2014
+ */
+@Plugin(PluginInterface)
+abstract class AbstractPlugin implements PluginInterface{
+}