+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Add dependency on guice-multibindings
+* Add custom bindings definition in installer through BindingInstaller marker interface
+* Add PluginInstaller: shortcut for multibindings mechanism
+
 ### 0.9.0 (2014-09-05)
 
 * Initial release